adjust embedding vector

+++ src/openai_leap.nim
@@ -34,7 +34,7 @@ type
     user*: Option[string]
   CreateEmbeddingRespObj* = ref object
     index*: int              # index into the input sequence in the request
-    embedding*: seq[float32] # https://platform.openai.com/docs/guides/embeddings
+    embedding*: seq[float64] # https://platform.openai.com/docs/guides/embeddings
     `object`*: string
   CreateEmbeddingResp* = ref object
     data*: seq[CreateEmbeddingRespObj]
